CSS rule contains two parts: a selector and a declaration.
Selectors indicate which element the rule applies to.

Declaration indicate how the element referred to in the selctor
should be styled.

Declarations are split into two parts property and value and are
seperated by colon and sit inside curly bracket after selectors

Selectors are of different types and are case-sensitive below are
some selectors

* {} - Universal Selector : Applies to all elements

h1 {} - Type Selector: Applies to element specified it can be any html
element

.class {} - Class Selector: Applies to element which belongs to the
specified class in its class attribute value

'#id {} - Id Selector: Applies to element which belongs to the specified
id in its id attribute value and can not be repeated again.

li>a {} - Child Selector: Applies to element which is direct child of
another element

p a {} - Descendant Selector:  Applies to element which is child of
another element not just a direct child.

h1+p {} - Adjacent Sibling Selector: Applies an element that is the
next sibling of another element

h1~p {} - General Sibling Selector: Applies an element that is a
sibling of another, although it does not have to be the directly
preceding element