CSS3 introduces the opacity property which allows you to specify the opacity of an element and any of its child elements.
The value is a number between 0.0 and 1.0 (so a value of 0.5 is 50%
opacity and 0.15 is 15% opacity).

The CSS3 rgba property allows you to specify a color, just like you
would with an RGB value, but adds a fourth value to indicate opacity.
This value is known as an alpha value and is a number between 0.0 and 1.0