The font-family property allows you to specify the typeface that should be used for any text inside the element(s) to which a CSS rule applies.
The user's computer needs the typeface installed. CSS is used to specify
the typeface.

font-size: The font-size property enables you to specify a size for the
font.

Three ways to define size of a font
Pixels - Pixels are commonly used unit because of precise control
Percentages - The default size of text in browsers is 16px. So a size of
75% would be the equivalent of 12px, and 200% would be 32px.
EMS - An em is equivalent to the width of a letter m.
EMS allow you to change the size of text relative to the size of the
text in the parent element.