font-weight: The font-weight property allows you to create bold text. options are: normal and bold.
font-style: If you want to create italic text, you can use the
font-style property. options are: normal, italic and oblique

text-transform: The text-transform property is used to change the case
of text giving it one of the following values:
uppercase - This causes the text to appear uppercase.
lowercase - This causes the text to appear lowercase.
capitalize - This causes the first letter of each word to appear
capitalized.

text-decoration: The text-decoration property allows you to specify the
following rules:
none - This removes any decoration already applied to the text.
underline - This adds a line underneath the text.
overline - This adds a line over the top of the text.
line-through -This adds a line through words.
blink - This animates the text to make it flash on and off

line-height: the line-height property sets the height of an entire line
of text, so the difference between the fontsize and the line-height is
equivalent to the leading; Leading is measured from the bottom of the
descender on one line to the top of the ascender on the next.

letter-spacing: Kerning is the term typographers use for the space
between each letter

word-spacing: the gap between words can be control by using the
word-spacing property

text-align: The text-align property allows you to control the alignment
of text. The property can take one of four values:
left, right, center and justify

vertical-align: It is more commonly used with inline elements such as
<img>, <em>, or <strong> elements.
baseline, sub, super, top, text-top, middle, bottom and text-bottom are
options for vertical-align property

text-indent: The text-indent property allows you to indent the first
line of text within an element.